Move theme.tsx out of pages folder to root
