+++ pages/_document.tsx
@@ -1,6 +1,6 @@
 import { ColorModeScript } from "@chakra-ui/react";
 import NextDocument, { Html, Head, Main, NextScript } from "next/document";
-import theme from "./theme";
+import theme from "../theme";
 
 export default class Document extends NextDocument {
   render() {
